+++ R/copy-to.r
@@ -1,4 +1,4 @@
-#' Copy a local data frame to a remote src.
+#' Copy a local data frame to a remote src
 #'
 #' This function uploads a local data frame into a remote data source, creating
 #' the table definition as needed. Wherever possible, the new object will be
@@ -25,7 +25,7 @@ copy_to <- function(dest, df, name = deparse(substitute(df)),
   UseMethod("copy_to")
 }
 
-#' Copy tables to same source, if necessary.
+#' Copy tables to same source, if necessary
 #'
 #' @param x,y `y` will be copied to `x`, if neccessary.
 #' @param copy If `x` and `y` are not from the same data source,