+++ src/init.cpp
@@ -1,3 +1,4 @@
+#include "pch.h"
 #include <dplyr/main.h>
 
 #include <dplyr/registration.h>
@@ -5,7 +6,7 @@
 using namespace Rcpp;
 using namespace dplyr;
 
-#define DPLYR_REGISTER(__FUN__) R_RegisterCCallable( "dplyr", #__FUN__, (DL_FUNC)__FUN__ );
+#define DPLYR_REGISTER(__FUN__) do { R_RegisterCCallable( "dplyr", #__FUN__, (DL_FUNC)__FUN__ ); } while(0);
 
 SEXP get_cache() {
   static SEXP cache = 0;
@@ -30,10 +31,10 @@ SEXP get_time_classes() {
 }
 
 extern "C" void R_init_dplyr(BOOST_ATTRIBUTE_UNUSED DllInfo* info) {
-  DPLYR_REGISTER(build_index_cpp)
-  DPLYR_REGISTER(registerHybridHandler)
+  DPLYR_REGISTER(build_index_cpp);
+  DPLYR_REGISTER(registerHybridHandler);
 
-  DPLYR_REGISTER(get_time_classes)
-  DPLYR_REGISTER(get_date_classes)
+  DPLYR_REGISTER(get_time_classes);
+  DPLYR_REGISTER(get_date_classes);
 }